apply clippy suggestions

+++ src/modules/utils/directory_nix.rs
@@ -23,13 +23,13 @@ pub fn is_write_allowed(folder_path: &Path) -> Result<bool, String> {
         return Ok(true);
     }
     if meta.uid() == euid.as_raw() {
-        Ok(perms & Mode::S_IWUSR.bits() as u32 != 0)
+        Ok(perms & Mode::S_IWUSR.bits() != 0)
     } else if (meta.gid() == Gid::effective().as_raw())
         || (get_supplementary_groups().contains(&meta.gid()))
     {
-        Ok(perms & Mode::S_IWGRP.bits() as u32 != 0)
+        Ok(perms & Mode::S_IWGRP.bits() != 0)
     } else {
-        Ok(perms & Mode::S_IWOTH.bits() as u32 != 0)
+        Ok(perms & Mode::S_IWOTH.bits() != 0)
     }
 }
